Reorganize setup components into legacy and setup folder.
