+++ assets/js/components/legacy-setup/site-verification.js
