+++ assets/js/components/legacy-setup/wizard-progress-step.js
