+++ assets/js/components/legacy-setup/wizard-step-authentication.js
