+++ assets/js/components/legacy-setup/wizard-step-complete-setup.js
