+++ assets/js/components/legacy-setup/wizard-step-search-console-property.js
@@ -29,7 +29,7 @@ import { Component } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import SearchConsole from '../../components/setup/search-console';
+import SearchConsole from '../../components/legacy-setup/search-console';
 
 class WizardStepSearchConsoleProperty extends Component {
 	render() {
