+++ assets/js/googlesitekit-module.js
@@ -35,7 +35,7 @@ import './components/data';
 import './components/notifications';
 import Root from './components/root';
 import ModuleApp from './components/module-app';
-import Setup from './components/setup/setup-wrapper';
+import Setup from './components/setup/ModuleSetup';
 
 class GoogleSitekitModule extends Component {
 	constructor( props ) {
